refactor(error): extract errors to module with DataType fields
